+++ src/App.css
@@ -40,3 +40,35 @@ pre {
   overflow-x: auto;
   max-height: 600px;
 }
+
+/* Chart Buttons Row Layout */
+.chart-buttons-container {
+  display: flex;
+  flex-direction: row;
+  gap: 20px;
+  margin: 30px 0;
+  padding: 20px;
+  background: linear-gradient(135deg, #f8f9fa 0%, #e9ecef 100%);
+  border-radius: 12px;
+  border: 1px solid #dee2e6;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  align-items: flex-start;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+/* Responsive behavior for smaller screens */
+@media (max-width: 768px) {
+  .chart-buttons-container {
+    flex-direction: column;
+    gap: 15px;
+    align-items: stretch;
+  }
+}
+
+/* Ensure button components fit well in the row */
+.chart-buttons-container > * {
+  flex: 1;
+  min-width: 280px;
+  max-width: 400px;
+}
