+++ src/App.jsx
@@ -18,17 +18,85 @@ function App() {
   const [dividendChart, setDividendChart] = useState(null);
   const [calendarChart, setCalendarChart] = useState(null);
   const [realizedGainsChart, setRealizedGainsChart] = useState(null);
+  const [activeChart, setActiveChart] = useState(null);
+  
+  // Refs to store reset functions for each button component
+  const dividendButtonRef = React.useRef(null);
+  const calendarButtonRef = React.useRef(null);
+  const realizedGainsButtonRef = React.useRef(null);
+
+  const resetAllCharts = () => {
+    setDividendChart(null);
+    setCalendarChart(null);
+    setRealizedGainsChart(null);
+    setActiveChart(null);
+  };
+
+  const resetOtherButtons = (excludeButton) => {
+    // Reset other button states by calling their reset functions
+    if (excludeButton !== 'dividend' && dividendButtonRef.current?.resetButton) {
+      dividendButtonRef.current.resetButton();
+    }
+    if (excludeButton !== 'calendar' && calendarButtonRef.current?.resetButton) {
+      calendarButtonRef.current.resetButton();
+    }
+    if (excludeButton !== 'realized-gains' && realizedGainsButtonRef.current?.resetButton) {
+      realizedGainsButtonRef.current.resetButton();
+    }
+  };
 
   const handleDividendChartReady = (chartInfo) => {
-    setDividendChart(chartInfo);
+    if (chartInfo) {
+      // Reset other charts and buttons when dividend chart is activated
+      setCalendarChart(null);
+      setRealizedGainsChart(null);
+      resetOtherButtons('dividend');
+      
+      setActiveChart("dividends");    
+      setDividendChart(chartInfo);
+    } else {
+      // If chartInfo is null, just clear the dividend chart
+      setDividendChart(null);
+      if (activeChart === "dividends") {
+        setActiveChart(null);
+      }
+    }
   };
 
   const handleCalendarChartReady = (chartInfo) => {
-    setCalendarChart(chartInfo);
+    if (chartInfo) {
+      // Reset other charts and buttons when calendar chart is activated
+      setDividendChart(null);
+      setRealizedGainsChart(null);
+      resetOtherButtons('calendar');
+      
+      setActiveChart("calendar");    
+      setCalendarChart(chartInfo);
+    } else {
+      // If chartInfo is null, just clear the calendar chart
+      setCalendarChart(null);
+      if (activeChart === "calendar") {
+        setActiveChart(null);
+      }
+    }
   };
 
   const handleRealizedGainsChartReady = (chartInfo) => {
-    setRealizedGainsChart(chartInfo);
+    if (chartInfo) {
+      // Reset other charts and buttons when realized gains chart is activated
+      setDividendChart(null);
+      setCalendarChart(null);
+      resetOtherButtons('realized-gains');
+      
+      setActiveChart("realized-gains");    
+      setRealizedGainsChart(chartInfo);
+    } else {
+      // If chartInfo is null, just clear the realized gains chart
+      setRealizedGainsChart(null);
+      if (activeChart === "realized-gains") {
+        setActiveChart(null);
+      }
+    }
   };
 
   return (
@@ -37,30 +105,41 @@ function App() {
       
       <Parser setSectionsData={setSectionsData} setTotals={setTotals} setTrades={setTrades} />
 
-      {/* Extracted Dividend Button */}
-      <DividendChartButton 
-        sectionsData={sectionsData}
-        onChartDataReady={handleDividendChartReady}
-      />
+      {/* Chart Buttons Row */}
+      <div className="chart-buttons-container">
+        {/* Extracted Dividend Button */}
+        <DividendChartButton 
+          ref={dividendButtonRef}
+          sectionsData={sectionsData}
+          onChartDataReady={handleDividendChartReady}
+        />
 
+        {/* Extracted Calendar Button */}
+        <CalendarChartButton 
+          ref={calendarButtonRef}
+          dateData={trades}
+          sectionsData={sectionsData}
+          buttonText="Calendar Chart"
+          defaultBoxColor="#f5f5f5"
+          boxBorderColor="#cccccc"
+          rowCount={3}
+          onChartDataReady={handleCalendarChartReady}
+        />      
+
+        {/* Extracted Realized Gains Buttons Group */}
+        <RealizedGainsButtonsGroup 
+          ref={realizedGainsButtonRef}
+          totals={totals}
+          sectionsData={sectionsData}
+          onChartDataReady={handleRealizedGainsChartReady}
+        />
+      </div>
       {/* Dividend Chart Display */}
-      {dividendChart && dividendChart.type === 'dividends' && (
+      {dividendChart && activeChart === 'dividends' && (          
         <DividendChartComponent sectionsData={sectionsData} />
       )}
-
-      {/* Extracted Calendar Button */}
-      <CalendarChartButton 
-        dateData={trades}
-        sectionsData={sectionsData}
-        buttonText="Calendar Chart"
-        defaultBoxColor="#f5f5f5"
-        boxBorderColor="#cccccc"
-        rowCount={3}
-        onChartDataReady={handleCalendarChartReady}
-      />
-
       {/* Calendar Chart Display */}
-      {calendarChart && calendarChart.type === 'calendar' && (
+      {calendarChart && activeChart === 'calendar' && (
         <CalendarChartComponent 
           dateData={trades}
           sectionsData={sectionsData}
@@ -70,15 +149,8 @@ function App() {
         />
       )}
 
-      {/* Extracted Realized Gains Buttons Group */}
-      <RealizedGainsButtonsGroup 
-        totals={totals}
-        sectionsData={sectionsData}
-        onChartDataReady={handleRealizedGainsChartReady}
-      />
-
       {/* Realized Gains Chart Display */}
-      {realizedGainsChart && realizedGainsChart.type === 'realized-gains' && (
+      {realizedGainsChart && activeChart === 'realized-gains' && (
         <RealizedGainsComponent 
           chartData={realizedGainsChart.data}
           selectedSources={realizedGainsChart.selectedSources || []}
